[FIX] Allow popup to open by fixing bing icon
When an favicon from a website respond with a timeout the popup window
of the extension is not opening until several seconds.
Here the bing icon was not available anymore, therefore the popup was
not opening instantly.

This is a temporary fix to that but ultimately the icons should
definitely be stored in local.