[REF] Style selector independant for each extension
Sometimes if we edit the css for an extension it is not applied if there is one of the other extension that still has the old css. Or worse the new css can be applied to the panels from the other extensions which can lead to bad styling.

Putting a separate parent selector to each css rule allows to make some updates on the css that will be taken into account only for the extension that is up-to-date.