[FIX] Premium as setting
In the content script, the premium is now managed as if it was a setting. This way we can apply event on it when we receive the information from extpay.

Settings event dispatchers have also consequently been improved so that anytime we change a setting it fires an event no matter where the change was made