[FIX] ChatGPT: put back event-stream
It seems like some users still have the old way of operating ofChatGPT
with EventStream. I thus need to put back the code from before
[this commit][1].

Now it should operate well no matter if it's a websocket or an
event-stream.

Although not sure if this solve [this issue][2] because I cannot
reproduce it in local.

I came to this conclusion because a user send me the HAR of the requests
made by OptiSearch and the /conversation request responds with a
`text/event-stream` content-type but for me it's a json with info about
the websocket.

[1]: https://github.com/dj0ulo/OptiSearch/commit/8d90d8d
[2]: https://github.com/Dj0ulo/OptiSearch/issues/37